test(config): split production helper and settings coverage
