+++ tests/test_production_settings.py
@@ -1,58 +1,9 @@
-"""Tests for production settings and environment parsing helpers."""
+"""Tests for production settings module behavior."""
 
 from __future__ import annotations
 
 import importlib
 import sys
-
-import pytest
-
-from config.settings.production_helpers import env_bool, env_int, env_list
-
-
-@pytest.mark.parametrize(
-    ("raw_value", "default", "expected"),
-    [
-        (None, False, False),
-        (None, True, True),
-        ("1", False, True),
-        ("true", False, True),
-        ("YES", False, True),
-        ("on", False, True),
-        ("0", True, False),
-        ("false", True, False),
-    ],
-)
-def test_env_bool_parses_truthy_falsey_and_default_values(
-    monkeypatch, raw_value: str | None, default: bool, expected: bool
-) -> None:
-    """Boolean helper should support explicit truthy values and fallback defaults."""
-    if raw_value is None:
-        monkeypatch.delenv("BOOL_SETTING", raising=False)
-    else:
-        monkeypatch.setenv("BOOL_SETTING", raw_value)
-
-    assert env_bool("BOOL_SETTING", default) is expected
-
-
-def test_env_int_parses_integer_values_and_defaults(monkeypatch) -> None:
-    """Integer helper should return defaults when unset and parse numeric strings."""
-    monkeypatch.delenv("INT_SETTING", raising=False)
-    assert env_int("INT_SETTING", 7) == 7
-
-    monkeypatch.setenv("INT_SETTING", "42")
-    assert env_int("INT_SETTING", 7) == 42
-
-
-def test_env_list_splits_comma_separated_values_and_handles_defaults(monkeypatch) -> None:
-    """List helper should trim values and preserve the provided default list."""
-    monkeypatch.delenv("LIST_SETTING", raising=False)
-    assert env_list("LIST_SETTING", ["a", "b"]) == ["a", "b"]
-    assert env_list("LIST_SETTING") == []
-
-    monkeypatch.setenv("LIST_SETTING", " alpha, beta ,, gamma ")
-    assert env_list("LIST_SETTING") == ["alpha", "beta", "gamma"]
-
 
 def test_production_settings_apply_env_driven_security_and_release_values(monkeypatch) -> None:
     """Production settings should expose the expected hardened env-driven values."""
